requirements.txt
azure-ai-formrecognizer==3.3.3
azure-identity==1.25.1
azure-storage-blob==12.27.1
backports.tarfile==1.2.0
fastapi==0.124.4
importlib-metadata==8.0.0
jaraco.collections==5.1.0
langchain==1.2.0
langchain-community==0.4.1
langchain-openai==1.1.4
num2words==0.5.14
pandas==2.3.3
pip-chill==1.0.3
platformdirs==4.2.2
pymongo==4.15.5
python-multipart==0.0.20
tomli==2.0.1
uvicorn==0.38.0
